feat:Implement header styling, icons, shared styles, and layout adjust
