+++ src/app/components/sharedStyles.tsx
@@ -1,0 +1,14 @@
+'use client'
+import styled from "styled-components";
+
+const StyledHeader = styled.header`
+  background-color: ${({ theme }) => theme.colors.background};
+  color: white;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  padding: 16px 20px 20px;
+`;
+
+export { StyledHeader };
