+++ src/app/page.tsx
@@ -1,9 +1,9 @@
-"use-client";
-
 export default function Home() {
   return (
-    <main>
-     <h1>Meus cursos</h1>
-    </main>
-  )
+    <>
+      <main>
+        <h1>Meus cursos</h1>
+      </main>
+    </>
+  );
 }
